% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvMF.R
\name{dvMFangle}
\alias{dvMFangle}
\title{Probability density function of angle between vMF mean direction and random vector}
\usage{
dvMFangle(r, p, kappa)
}
\arguments{
\item{r}{vector of quantiles. \eqn{-1\le r\le 1}.}

\item{p}{dimension of the vMF distribution. i.e., vMF on \eqn{S^{p-1}}, \eqn{p\ge 2}.}

\item{kappa}{concentration parameter. \eqn{kappa\ge 0}.}
}
\value{
vector of density.
}
\description{
Probability density function of angle between vMF mean direction and random vector
}
\examples{
dvMFangle(seq(-1,1,by=0.01), 2, 10)
dvMFangle(seq(0,1,by=0.01), 3, 0)
}
