% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvMF-package.R
\docType{package}
\name{rvMF-package}
\alias{rvMF-package}
\alias{_PACKAGE}
\title{rvMF: Fast Generation of von Mises-Fisher Distributed Random Vectors}
\description{
Fast generation of von Mises-Fisher distributed random vectors. References to be added.
}
\author{
\strong{Maintainer}: Seungwoo Kang \email{kangsw0401@snu.ac.kr}

Authors:
\itemize{
  \item Hee-Seok Oh
}

}
\keyword{internal}
