% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvMF.R
\name{rvMF}
\alias{rvMF}
\title{Fast random vectors generator from a von Mises-Fisher distribution}
\usage{
rvMF(n, mu, k)
}
\arguments{
\item{n}{number of random vectors to generate.}

\item{mu}{mean direction.}

\item{k}{concentration parameter. \eqn{k\ge 0}.}
}
\value{
matrix where each row follows iid vMF distribution.
}
\description{
Fast random vectors generator from a von Mises-Fisher distribution
}
\examples{
rvMF(1e5, c(0,0,1), 10)
rvMF(1e4, c(1,1)/sqrt(2), 0)
}
