% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvMF.R
\name{rvMFangle}
\alias{rvMFangle}
\title{Generate random variable following angle between vMF mean direction and random vector}
\usage{
rvMFangle(n, p, kappa)
}
\arguments{
\item{n}{number of random vectors to generate.}

\item{p}{dimension of the vMF distribution. i.e., vMF on \eqn{S^{p-1}}, \eqn{p\ge 2}.}

\item{kappa}{concentration parameter. \eqn{kappa\ge 0}.}
}
\value{
vector where each component follows iid vMF angle distribution.
}
\description{
Generate random variable following angle between vMF mean direction and random vector
}
\examples{
rvMFangle(1e5, 2, 10)
rvMFangle(1e4, 3, 0)
}
